updated return value of getFssIndex from SrcIdx; double to int

+++ dev/scratch/UCERF3/erf/FaultSystemSolutionPoissonERF.java
@@ -370,7 +370,7 @@ public class FaultSystemSolutionPoissonERF extends AbstractERF {
 	 * @param iSrc
 	 * @return
 	 */
-	public double getFltSysRupIndexForSource(int iSrc) {
+	public int getFltSysRupIndexForSource(int iSrc) {
 		return fltSysRupIndexForSource[iSrc];
 	}
